// src/integrations/supabase/types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.3 (519615d)"
  }
  public: {
    Tables: {
// ...
      media_items: {
        Row: {
          blur_settings: Json | null
          created_at: string
          description: string | null
          external_link: string | null
          hover_unblur: boolean | null
          id: string
          is_blurred: boolean | null
          is_locked: boolean | null
          is_main: boolean | null
          link: string | null
          name: string | null
          pinned: boolean | null
          price: string | null
          storage_path: string
          timer_settings: Json | null
          type: string
          updated_at: string
          user_id: string
        }
        Insert: {
          blur_settings?: Json | null
          created_at?: string
          description?: string | null
          external_link?: string | null
          hover_unblur?: boolean | null
          id?: string
          is_blurred?: boolean | null
          is_locked?: boolean | null
          is_main?: boolean | null
          link?: string | null
          name?: string | null
          pinned?: boolean | null
          price?: string | null
          storage_path: string
          timer_settings?: Json | null
          type: string
          updated_at?: string
          user_id: string
        }
        Update: {
          blur_settings?: Json | null
          created_at?: string
          description?: string | null
          external_link?: string | null
          hover_unblur?: boolean | null
          id?: string
          is_blurred?: boolean | null
          is_locked?: boolean | null
          is_main?: boolean | null
          link?: string | null
          name?: string | null
          pinned?: boolean | null
          price?: string | null
          storage_path?: string
          timer_settings?: Json | null
          type?: string
          updated_at?: string
          user_id?: string
        }
        Relationships: []
      }
// ...
      vitrine: {
        Row: {
          created_at: string | null
          id: string
          media: Json | null
          owner_id: string
          updated_at: string | null
          visible: boolean | null
        }
// ...
      }
// ...
    }
// ...
  }
}